style: remove shared folder and move shared files to src/dappsExplorer
