+++ src/dappsExplorer/filter/DAppsExplorerScreenFilter.tsx
@@ -25,13 +25,13 @@ import {
 } from 'src/dapps/selectors'
 import { fetchDappsList } from 'src/dapps/slice'
 import { DappSection, DappV1, DappV2 } from 'src/dapps/types'
+import DappCard from 'src/dappsExplorer/DappCard'
+import DappFilterChip from 'src/dappsExplorer/DappFilterChip'
 import FavoriteDappsSection from 'src/dappsExplorer/filter/FavoriteDappsSection'
 import { NoResults } from 'src/dappsExplorer/filter/NoResults'
-import DappCard from 'src/dappsExplorer/shared/DappCard'
-import DappFilterChip from 'src/dappsExplorer/shared/DappFilterChip'
-import useDappFavoritedToast from 'src/dappsExplorer/shared/useDappFavoritedToast'
-import useDappInfoBottomSheet from 'src/dappsExplorer/shared/useDappInfoBottomSheet'
-import useOpenDapp from 'src/dappsExplorer/shared/useOpenDapp'
+import useDappFavoritedToast from 'src/dappsExplorer/useDappFavoritedToast'
+import useDappInfoBottomSheet from 'src/dappsExplorer/useDappInfoBottomSheet'
+import useOpenDapp from 'src/dappsExplorer/useOpenDapp'
 import { currentLanguageSelector } from 'src/i18n/selectors'
 import Help from 'src/icons/Help'
 import DrawerTopBar from 'src/navigator/DrawerTopBar'
