+++ src/dappsExplorer/filter/FavoriteDappsSection.tsx
@@ -4,8 +4,8 @@ import { StyleSheet, Text, View } from 'react-native'
 import { useSelector } from 'react-redux'
 import { favoriteDappsSelector } from 'src/dapps/selectors'
 import { ActiveDapp, DappSection, isDappV2 } from 'src/dapps/types'
+import DappCard from 'src/dappsExplorer/DappCard'
 import { NoResults } from 'src/dappsExplorer/filter/NoResults'
-import DappCard from 'src/dappsExplorer/shared/DappCard'
 import StarIllustration from 'src/icons/StarIllustration'
 import Colors from 'src/styles/colors'
 import fontStyles from 'src/styles/fonts'
