+++ src/dappsExplorer/legacy/DAppsExplorerScreenLegacy.tsx
@@ -26,12 +26,12 @@ import {
 } from 'src/dapps/selectors'
 import { fetchDappsList } from 'src/dapps/slice'
 import { DappSection, DappV1, isDappV2 } from 'src/dapps/types'
+import DappCard from 'src/dappsExplorer/DappCard'
 import FavoriteDappsSection from 'src/dappsExplorer/legacy/FavoriteDappsSection'
 import FeaturedDappCard from 'src/dappsExplorer/legacy/FeaturedDappCard'
-import DappCard from 'src/dappsExplorer/shared/DappCard'
-import useDappFavoritedToast from 'src/dappsExplorer/shared/useDappFavoritedToast'
-import useDappInfoBottomSheet from 'src/dappsExplorer/shared/useDappInfoBottomSheet'
-import useOpenDapp from 'src/dappsExplorer/shared/useOpenDapp'
+import useDappFavoritedToast from 'src/dappsExplorer/useDappFavoritedToast'
+import useDappInfoBottomSheet from 'src/dappsExplorer/useDappInfoBottomSheet'
+import useOpenDapp from 'src/dappsExplorer/useOpenDapp'
 import Help from 'src/icons/Help'
 import { dappListLogo } from 'src/images/Images'
 import DrawerTopBar from 'src/navigator/DrawerTopBar'
