+++ src/dappsExplorer/legacy/FavoriteDappsSection.tsx
@@ -4,7 +4,7 @@ import { StyleSheet, Text, View } from 'react-native'
 import { useSelector } from 'react-redux'
 import { favoriteDappsSelector } from 'src/dapps/selectors'
 import { ActiveDapp, DappSection } from 'src/dapps/types'
-import DappCard from 'src/dappsExplorer/shared/DappCard'
+import DappCard from 'src/dappsExplorer/DappCard'
 import StarIllustration from 'src/icons/StarIllustration'
 import Colors from 'src/styles/colors'
 import fontStyles from 'src/styles/fonts'
