+++ src/dappsExplorer/useOpenDapp.tsx
@@ -6,7 +6,7 @@ import { activeScreenSelector } from 'src/app/selectors'
 import { dappsMinimalDisclaimerEnabledSelector, recentDappsSelector } from 'src/dapps/selectors'
 import { dappSelected } from 'src/dapps/slice'
 import { ActiveDapp, isDappV2 } from 'src/dapps/types'
-import DAppsBottomSheet from 'src/dappsExplorer/shared/DAppsBottomSheet'
+import DAppsBottomSheet from 'src/dappsExplorer/DAppsBottomSheet'
 import { Screens } from 'src/navigator/Screens'
 import { isDeepLink } from 'src/utils/linking'
 import Logger from 'src/utils/Logger'
